Split dynamic exchange behaviour in orderCreators and candleProcessors
