+++ src/sockTrader/core/exchanges/utils/orderTracker.ts
@@ -1,20 +1,45 @@
-import {IOrder} from "../../types/order";
+import {IOrder, OrderStatus, ReportType} from "../../types/order";
 
 export default class OrderTracker {
 
     private readonly unconfirmedOrders: Record<string, boolean> = {};
     private openOrders: IOrder[] = [];
 
-    setOrderUnconfirmed(orderId: string) {
-        this.unconfirmedOrders[orderId] = true;
+    private setOrderConfirmed(orderId: string) {
+        delete this.unconfirmedOrders[orderId];
+    }
+
+    private replaceOpenOrder(newOrder: IOrder, oldOrderId: string): IOrder | undefined {
+        const oldOrder = this.findOpenOrder(oldOrderId);
+        this.setOrderConfirmed(oldOrderId);
+        this.removeOpenOrder(oldOrderId);
+        this.addOpenOrder(newOrder);
+
+        return oldOrder;
+    }
+
+    private addOpenOrder(order: IOrder) {
+        this.openOrders.push(order);
+    }
+
+    private removeOpenOrder(orderId: string) {
+        this.openOrders = this.openOrders.filter(o => o.id !== orderId);
+    }
+
+    private findOpenOrder(orderId: string) {
+        return this.openOrders.find(openOrder => openOrder.id === orderId);
+    }
+
+    isOrderConfirmed(orderId: string) {
+        return !this.isOrderUnconfirmed(orderId);
     }
 
     isOrderUnconfirmed(orderId: string) {
         return this.unconfirmedOrders[orderId];
     }
 
-    setOrderConfirmed(orderId: string) {
-        delete this.unconfirmedOrders[orderId];
+    setOrderUnconfirmed(orderId: string) {
+        this.unconfirmedOrders[orderId] = true;
     }
 
     getOpenOrders() {
@@ -25,40 +50,26 @@ export default class OrderTracker {
         this.openOrders = orders;
     }
 
-    replaceOpenOrder(newOrder: IOrder, oldOrderId: string): IOrder | undefined {
-        const oldOrder = this.findOpenOrder(oldOrderId);
-        this.setOrderConfirmed(oldOrderId);
-        this.removeOpenOrder(oldOrderId);
-        this.addOpenOrder(newOrder);
-
-        return oldOrder;
-    }
-
-    addOpenOrder(order: IOrder) {
-        this.openOrders.push(order);
-    }
-
-    removeOpenOrder(orderId: string) {
-        this.openOrders = this.openOrders.filter(o => o.id !== orderId);
-    }
-
-    findOpenOrder(orderId: string) {
-        return this.openOrders.find(openOrder => openOrder.id === orderId);
-    }
-
     /**
-     * Validates if you can adjust an existing order on an exchange
-     * @param order the order to check
-     * @param price new price
-     * @param qty new quantity
+     * Processes order depending on the reportType
+     * @param order
      */
-    canAdjustOrder(order: IOrder, price: number, qty: number): boolean {
-        if (this.isOrderUnconfirmed(order.id)) return false;
+    process(order: IOrder) {
+        const orderId = order.id;
+        let oldOrder: IOrder | undefined;
 
-        // No need to replace!
-        if (order.price === price && order.quantity === qty) return false;
+        this.setOrderConfirmed(orderId);
 
-        this.setOrderUnconfirmed(order.id);
-        return true;
+        if (order.reportType === ReportType.REPLACED && order.originalId) {
+            oldOrder = this.replaceOpenOrder(order, order.originalId);
+        } else if (order.reportType === ReportType.NEW) {
+            this.addOpenOrder(order); // New order created
+        } else if (order.reportType === ReportType.TRADE && order.status === OrderStatus.FILLED) {
+            this.removeOpenOrder(orderId); // Order is 100% filled
+        } else if ([ReportType.CANCELED, ReportType.EXPIRED, ReportType.SUSPENDED].indexOf(order.reportType) > -1) {
+            this.removeOpenOrder(orderId); // Order is invalid
+        }
+
+        return {order, oldOrder};
     }
 }
